LayoutTests/webrtc/webrtc-transform.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <title>Testing media exchange with transform</title>
        <script src="../resources/testharness.js"></script>
        <script src="../resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id="video1" autoplay></video>
        <video id="video2" autoplay></video>
        <script src ="routines.js"></script>
        <script>
promise_test(async (test) => {
    if (!window.internals)
        return;

    let sender, receiver;
    let senderTransform = internals.createMockRTCRtpTransform();
    let receiverTransform = internals.createMockRTCRtpTransform();

    const localStream = await navigator.mediaDevices.getUserMedia({audio: true});
    const stream = await new Promise((resolve, reject) => {
        createConnections((firstConnection) => {
            pc1 = firstConnection;
            sender = firstConnection.addTrack(localStream.getAudioTracks()[0], localStream);
            sender.transform = senderTransform;
        }, (secondConnection) => {
            pc2 = secondConnection;
            secondConnection.ontrack = (trackEvent) => {
                receiver = trackEvent.receiver;
                receiver.transform = receiverTransform;
                resolve(trackEvent.streams[0]);
            };
        });
        setTimeout(() => reject("Test timed out"), 5000);
    });

    video1.srcObject = stream;
    await video1.play();

    let counter = 0;
    while (!receiverTransform.isProcessing && counter++ < 100)
        await new Promise(resolve => setTimeout(resolve, 50));

    assert_true(senderTransform.isProcessing, "sender");
}, "audio exchange with transform");

promise_test(async (test) => {
    if (!window.internals)
        return;

    let sender, receiver;
    let senderTransform = internals.createMockRTCRtpTransform();
    let receiverTransform = internals.createMockRTCRtpTransform();

    const localStream = await navigator.mediaDevices.getUserMedia({video: true});

    const stream = await new Promise((resolve, reject) => {
        createConnections((firstConnection) => {
            pc1 = firstConnection;
            sender = firstConnection.addTrack(localStream.getVideoTracks()[0], localStream);
            sender.transform = senderTransform;
        }, (secondConnection) => {
            pc2 = secondConnection;
            secondConnection.ontrack = (trackEvent) => {
                receiver = trackEvent.receiver;
                receiver.transform = receiverTransform;
                resolve(trackEvent.streams[0]);
            };
        });
        setTimeout(() => reject("Test timed out"), 5000);
    });

    video2.srcObject = stream;
    await video2.play();

   let counter = 0;
   while (!receiverTransform.isProcessing && counter++ < 100)
       await new Promise(resolve => setTimeout(resolve, 50));

    assert_true(senderTransform.isProcessing, "sender");
}, "video exchange with transform");
        </script>
    </body>
</html>
